feat: unified open_s3 API, S3Credentials, object_store 0.14
- Upgrade object_store 0.11 → 0.14 (aws-lc-rs replaces ring, fixes aarch64 CI)
- Add S3Credentials struct for caller-owned credential management
- Unified open_s3(remote, budget, cache_dir): cache_dir presence decides RAM vs disk cache
- Configurable BlockCache RAM cap via CacheBudget
- Python wrapper exposes open_s3 staticmethod with --features s3
- README: add S3 remote storage, search index, transaction docs
- S3 full parity tests: 15 tests covering SQL, MATCH, SHORTEST, GROUP BY, disk cache

+++ src/engine/cache.rs
@@ -13,7 +13,7 @@ use std::path::PathBuf;
 use std::sync::Arc;
 
 use object_store::path::Path as ObjPath;
-use object_store::ObjectStore;
+use object_store::{ObjectStore, ObjectStoreExt};
 use tokio::runtime::Runtime;
 
 pub const BLOCK_SIZE: usize = 64 * 1024; // 64 KB
@@ -194,7 +194,7 @@ impl BlockCache {
 
         let data = self.runtime.block_on(async {
             self.store
-                .get_range(&self.obj_path, block_start as usize..block_end)
+                .get_range(&self.obj_path, block_start..block_end as u64)
                 .await
                 .ok()
         })?;
